refactor: move dialog to shared-ui (#1005)
