+++ apps/journeys-admin/src/components/JourneyList/ActiveJourneyList/ActiveJourneyList.tsx
@@ -2,6 +2,7 @@ import { ReactElement, useEffect, useState } from 'react'
 import Card from '@mui/material/Card'
 import { gql, useMutation, useQuery } from '@apollo/client'
 import Typography from '@mui/material/Typography'
+import { Dialog } from '@core/shared/ui/Dialog'
 import { useTranslation } from 'react-i18next'
 import { AuthUser } from 'next-firebase-auth'
 import { useSnackbar } from 'notistack'
@@ -12,7 +13,6 @@ import {
 import { JourneyCard } from '../JourneyCard'
 import { AddJourneyButton } from '../AddJourneyButton'
 import { SortOrder } from '../JourneySort'
-import { Dialog } from '../../Dialog'
 import { sortJourneys } from '../JourneySort/utils/sortJourneys'
 import { getDuplicatedJourney } from './utils/getDuplicatedJourney'
