+++ apps/journeys-admin/src/components/JourneyList/JourneyCard/JourneyCardMenu/DeleteJourneyDialog/DeleteJourneyDialog.tsx
@@ -1,10 +1,10 @@
 import { ReactElement } from 'react'
 import { useMutation, gql, ApolloQueryResult } from '@apollo/client'
 import Typography from '@mui/material/Typography'
+import { Dialog } from '@core/shared/ui/Dialog'
 import { useSnackbar } from 'notistack'
 import { JourneyDelete } from '../../../../../../__generated__/JourneyDelete'
 import { JourneyStatus } from '../../../../../../__generated__/globalTypes'
-import { Dialog } from '../../../../Dialog'
 import { GetActiveJourneys } from '../../../../../../__generated__/GetActiveJourneys'
 import { GetArchivedJourneys } from '../../../../../../__generated__/GetArchivedJourneys'
 import { GetTrashedJourneys } from '../../../../../../__generated__/GetTrashedJourneys'
