+++ apps/journeys-admin/src/components/JourneyList/JourneyCard/JourneyCardMenu/RestoreJourneyDialog/RestoreJourneyDialog.tsx
@@ -2,9 +2,9 @@ import { ReactElement } from 'react'
 import { useMutation, gql, ApolloQueryResult } from '@apollo/client'
 import { useSnackbar } from 'notistack'
 import Typography from '@mui/material/Typography'
+import { Dialog } from '@core/shared/ui/Dialog'
 import { JourneyRestore } from '../../../../../../__generated__/JourneyRestore'
 import { JourneyStatus } from '../../../../../../__generated__/globalTypes'
-import { Dialog } from '../../../../Dialog'
 import { GetActiveJourneys } from '../../../../../../__generated__/GetActiveJourneys'
 import { GetArchivedJourneys } from '../../../../../../__generated__/GetArchivedJourneys'
 import { GetTrashedJourneys } from '../../../../../../__generated__/GetTrashedJourneys'
