+++ apps/journeys-admin/src/components/JourneyView/Menu/LanguageDialog/LanguageDialog.tsx
@@ -2,9 +2,9 @@ import { ReactElement } from 'react'
 import { useMutation, useQuery, gql } from '@apollo/client'
 import { useSnackbar } from 'notistack'
 import { Formik, Form, FormikValues } from 'formik'
+import { Dialog } from '@core/shared/ui/Dialog'
 import { useJourney } from '@core/journeys/ui/JourneyProvider'
 import { JourneyLanguageUpdate } from '../../../../../__generated__/JourneyLanguageUpdate'
-import { Dialog } from '../../../Dialog'
 import { LanguageSelect } from '../../../LanguageSelect'
 import { GetLanguages } from '../../../../../__generated__/GetLanguages'
