+++ apps/journeys-admin/src/components/TemplateList/ActiveTemplates/ActiveTemplates.tsx
@@ -1,12 +1,12 @@
 import Typography from '@mui/material/Typography'
 import { ReactElement, useState, useEffect } from 'react'
 import { gql, useMutation, useQuery } from '@apollo/client'
+import { Dialog } from '@core/shared/ui/Dialog'
 import { AuthUser } from 'next-firebase-auth'
 import { useSnackbar } from 'notistack'
 import Card from '@mui/material/Card'
 import { useTranslation } from 'react-i18next'
 import { SortOrder } from '../../JourneyList/JourneySort'
-import { Dialog } from '../../Dialog'
 import { sortJourneys } from '../../JourneyList/JourneySort/utils/sortJourneys'
 import {
   ARCHIVE_ACTIVE_JOURNEYS,
